refactor(bootstrap): move Bootstrap.scala to the hydrozoa.bootstrap package
Colocate the config-authoring CLIs (Bootstrap, GenerateKeyPair, Migrate,
BuildHeadConfig) with InitializationTxBuilder in hydrozoa.bootstrap. Logger
names and runMain docstrings follow the new package; NodeConfig doc link
updated.
